Prezista Drug Label Data

A dataset containing an excerpt from the official Prezista FDA Drug Label to help in the app exploration.

prezista

Format

A dataset with 33 rows and 6 variables

covname

Covariate Name, a character variable with two values Protease Inihibitors and Other Antiretrovirals

label

Covariate value label, a character variable with several possible values

paramname

Parameter on which the effects are shown, a character variable with three possible values Cmax, AUC and Cmin

mid

Middle value for the effects, the median from the uncertainty distribution

lower

Lower value for the effects usually the 5% from the uncertainty distribution

upper

Upper value for the effects usually the 95% from the uncertainty distribution

Source

Table 16 from https://www.accessdata.fda.gov/drugsatfda_docs/label/2017/021976s045_202895s020lbl.pdf